refactor: Rename nav-related files
