+++ src/components/layout/header/Header.tsx
@@ -2,15 +2,14 @@
 
 import { MouseEvent } from "react";
 
-import NavToggleButton from "@/components/layout/nav/NavToggleButton";
-import NavOverlay from "@/components/layout/nav/NavOverlay";
-import NavPanel from "@/components/layout/nav/NavPanel";
-import ThemeToggle from "@/components/ui/ThemeToggle";
-import { useNavPanel } from "@/hooks/animations/useNavPanel";
+import { useToggleSidebar } from "@/hooks/animations/useToggleSidebar";
 import { navLinks } from "@/lib/site";
+import Sidebar from "@/components/layout/header/Sidebar";
+import SidebarToggle from "@/components/layout/header/Toggle";
+import ThemeToggle from "@/components/ui/ThemeToggle";
 
 export default function Nav() {
-  const { open, visible, toggle, close, panelRef, linksRef, iconsRef } = useNavPanel();
+  const { open, visible, toggle, close, panelRef, linksRef, iconsRef } = useToggleSidebar();
   const showNavToggle = navLinks.length > 0;
 
   const handleToggle = (event: MouseEvent<HTMLButtonElement>) => {
@@ -23,7 +22,7 @@ export default function Nav() {
       <header className="fixed top-0 left-0 right-0 z-50 flex items-center justify-between px-4 py-4 pointer-events-none">
         {/* Sidebar Toggle Button */}
         <div className="pointer-events-auto -ml-1">
-          {showNavToggle && <NavToggleButton open={open} onToggle={handleToggle} />}
+          {showNavToggle && <SidebarToggle open={open} onToggle={handleToggle} />}
         </div>
 
         {/* Theme Toggle */}
@@ -35,8 +34,16 @@ export default function Nav() {
       {/* Sidebar Panel */}
       {showNavToggle && (
         <>
-          <NavOverlay visible={visible} onClick={close} />
-          <NavPanel
+          {/* Overlay */}
+          {visible && (
+            <div
+              onClick={close}
+              className="fixed inset-0 z-30 bg-black/20 dark:bg-black/40"
+            />
+          )}
+
+          {/* Sidebar Panel */}
+          <Sidebar
             visible={visible}
             panelRef={panelRef}
             linksRef={linksRef}
