--- src/components/layout/nav/NavOverlay.tsx
@@ -1,18 +1,0 @@
-"use client";
-
-type NavOverlayProps = {
-  visible: boolean;
-  onClick: () => void;
-};
-
-export default function NavOverlay({ visible, onClick }: NavOverlayProps) {
-  if (!visible) return null;
-
-  return (
-    <div
-      onClick={onClick}
-      className="fixed inset-0 z-30 bg-black/20 dark:bg-black/40"
-    />
-  );
-}
-
